requirements.txt
streamlit 
google-cloud-speech
google-cloud-aiplatform
